requirements.txt
# Deep Learning Framework
tensorflow>=2.15.0
keras>=3.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Image Augmentation
imgaug>=0.4.0
albumentations>=1.1.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
pyyaml>=6.0
h5py>=3.1.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipython>=7.0.0
notebook>=6.0.0

# Model Visualization
graphviz>=0.17.0
pydot>=1.4.0

# Logging and Monitoring
tensorboard>=2.7.0
wandb>=0.12.0

# Testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Development
black>=21.0.0
flake8>=4.0.0
pre-commit>=2.15.0
